backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# File handling
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
